skills/monorepo/resources/clean-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
# Production server
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
jsonschema==4.20.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0

# Cache
redis==5.0.1

# Authentication
PyJWT==2.8.0